+++ src/components/Footers/footersStyling.css
@@ -1,0 +1,82 @@
+/* footer */
+
+.footers {
+  display: flex;
+  margin-top: 100px;
+  align-items: flex-start;
+  padding: 30px;
+  width: 100%;
+}
+.logo {
+  display: flex;
+  flex-direction: column;
+  gap: 40px;
+  width: 50%;
+}
+.logo-img {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+.design > h3,
+.design > span {
+  color: #000;
+  font-family: Nunito;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+
+.contact{
+  width: 25%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
+.contact > h2 {
+  color: #000;
+  font-family: Nunito;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 100%;
+}
+.contact > p {
+  color: #000;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 130%;
+  margin-left: 25px;
+  position: relative;
+}
+.contact > p > img {
+  position: absolute;
+  top: 0;
+  left: -25px;
+}
+
+.social{
+  display: flex;
+  flex-direction: column;
+ 
+  width: 25%;
+  align-items: flex-start;
+  justify-content: flex-start;
+}
+
+.social>h2{
+  color: #000;
+  font-family: Nunito;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 100%;
+}
+.social-img {
+  display: flex;
+  gap: 20px;
+}
