+++ src/components/ModalWindow/modalWindowStyling.css
@@ -1,0 +1,135 @@
+/* modal */
+
+.modal-container {
+  height: 100lvw;
+  width: 98.7vw;
+  position: absolute;
+  top: 0;
+  left: 0;
+  background-color: rgba(0, 0, 0, 0.75);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.modal {
+  width: fit-content;
+  height: fit-content;
+  background-color: #fff;
+  border-radius: 10px;
+  position: fixed;
+  top: 20%;
+  padding: 30px;
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  flex-direction: column;
+  gap: 30px;
+}
+.modal > h1 {
+  color: #000;
+  font-family: Nunito;
+  font-size: 40px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 120%;
+}
+.modal-info {
+  display: flex;
+  flex-direction: row;
+  gap: 50px;
+}
+.modal-info > img {
+  width: 250px;
+  height: 200px;
+}
+.modal-subinfo {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 344px;
+}
+.modal-subinfo > p {
+  color: #000;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 130%;
+}
+.modal-ingredients > h3 {
+  color: #000;
+  font-family: Nunito;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 130%;
+}
+.modal-ingredients > ul > li {
+  color: #000;
+  font-family: Nunito;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 130%;
+  list-style: none;
+}
+.modal-ingredients > span {
+  color: #b1b1b1;
+  font-family: Nunito;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 130%;
+}
+.modal-button {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  gap: 30px;
+}
+.modal-button > p {
+  color: #000;
+  font-family: Nunito;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 130%;
+}
+.modal-button > button {
+  width: 45%;
+  display: flex;
+  border-radius: 12px;
+  background: #ff7020;
+  height: 40px;
+  padding: 11px 32px;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 10px;
+  border: none;
+  cursor: pointer;
+  flex-shrink: 0;
+}
+.modal-count {
+  display: flex;
+  gap: 10px;
+  width: 10%;
+}
+.modal-count > button {
+  color: #000;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+  border: none;
+  cursor: pointer;
+}
+.modal-count > span {
+  color: #000;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
