+++ src/components/Products/productStyling.css
@@ -1,0 +1,89 @@
+/* products */
+
+.products-container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  padding: 15px;
+  width: 65%;
+}
+.products-container>h1 {
+  width: fit-content;
+  padding: 10px;
+  color:#FF7020;
+  font-family: Nunito;
+  font-size: 28px;
+  font-style: normal;
+  font-weight: 800;
+  letter-spacing: 5px;
+  border-bottom: #FFAB08 solid 5px;
+  border-radius: 50%;
+}
+.products {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 30px;
+}
+.product-item {
+  width: 250px;
+  height: 450px;
+  display: flex;
+  flex-direction: column;
+  /* align-items: center; */
+  justify-content: center;
+  gap: 20px;
+}
+.product-item > img {
+  width: 230px;
+}
+.product-item > h2 {
+  color: #000;
+  font-family: Nunito;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 100%;
+  justify-content: flex-start;
+  align-items: center;
+  display: flex;
+}
+.product-item > h3 {
+  color: #000;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 130%;
+  justify-content: flex-start;
+  align-items: center;
+  display: flex;
+}
+.product-item > span {
+  color: #b1b1b1;
+  font-family: Nunito;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 130%;
+  justify-content: flex-start;
+  align-items: center;
+  display: flex;
+}
+.product-item > button {
+  display: flex;
+  width: 236px;
+  height: 40px;
+  padding: 11px 82px;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  flex-shrink: 0;
+  border: none;
+  border-radius: 12px;
+  background: #aeaeb17e;
+  cursor: pointer;
+}
+.product-item > button:hover {
+  cursor: pointer;
+  background: #aeaeb1da;
+}
