+++ src/data/categories.js
@@ -2,46 +2,46 @@ export const categories = [
   {
     title: "Burger",
     id: "sku_id_1",
-    img:"./img/pic/1.png"
+    img:"./img/pic/1.svg"
   },
   {
     title: "Snacks",
     id: "sku_id_2",
-    img:"./img/pic/2.png"
+    img:"./img/pic/2.svg"
   },
   {
     title: "Hot-Dogs",
     id: "sku_id_3",
-    img:"./img/pic/3.png"
+    img:"./img/pic/3.svg"
   },
   {
     title: "Kombo",
     id: "sku_id_4",
-    img:"./img/pic/4.png"
+    img:"./img/pic/4.svg"
   },
   {
     title: "Shawarma",
     id: "sku_id_5",
-    img:"./img/pic/5.png"
+    img:"./img/pic/5.svg"
   },
   {
     title: "Pizza",
     id: "sku_id_6",
-    img:"./img/pic/6.png"
+    img:"./img/pic/6.svg"
   },
   {
     title: "Wok",
     id: "sku_id_7",
-    img:"./img/pic/7.png"
+    img:"./img/pic/7.svg"
   },
   {
     title: "Desserts",
     id: "sku_id_8",
-    img:"./img/pic/8.png"
+    img:"./img/pic/8.svg"
   },
   {
     title: "Sauces",
     id: "sku_id_9",
-    img:"./img/pic/9.png"
+    img:"./img/pic/9.svg"
   },
 ];
