+++ src/features/cart/cartSlice.js
@@ -2,7 +2,7 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
   cartItems: [],
-  quantity:1,
+  quantity:0,
   total: 0,
 };
